% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mloglikelihood.R
\name{Mloglikelihood}
\alias{Mloglikelihood}
\title{Title}
\usage{
Mloglikelihood(p, obs, dist, frailty, correct, transform = TRUE)
}
\arguments{
\item{p}{}

\item{obs}{}

\item{dist}{}

\item{frailty}{}

\item{correct}{}

\item{transform}{}
}
\description{
Title
}
